add bored_chat (WIP)
I was bored, and made a chat thing that looks like it could work peer to peer.
It's not like I know if it actually would, but now I can sleep hopefully.
Why didn't I just read the book, you ask.
Because I am physically not in reach of it.